* commit 'ad3002dbeee4b5c109f96a805fe69a63ec76a386': (56 commits)
  Use std::vector clear and push_back instead of resize in an effort to minimize memory allocations.
  Take advantage of the recent change where m_schema is correct for scalar and string to simplify declareShape to use no arguments.
  Fix datagroup destroy view and data to not destroy buffers when views are still attached to it.
  Restore const to isApplyValid and getVoidPtr signatures.  Add const_cast to conduit element_ptr call.
  Update setScalar, setString to update view schema.
  Expand isAllocated - ATK-566
  Remove use of F_string_result_as_arg option. Instead define MAXNAMESIZE=128.  Assume view and group names will not exceed 128 characters. ATK-540
  Fix the realloc test. Using getSchema in declareShape gets inconsistent values: after declare, before apply.  Change declareShape overload to use an explicit num_elems instead of extracting length from m_schema or m_node.
  Set shape for scalar and string.
  Add a url
  Setup a working example of breathe
  Change to the read-the-docs sphinx theme.
  Make paths in the top level web page relative. This allows a local install to work correctly.
  Rename ENABLE_PYTHON_MODULE to ENABLE_PYTHON_API ATK-469
  Change DataView.m_shape to be a std::vector instead of a pointer to a std::vector.  It always exists and has the shape, including 1-d. Before m_shape was only allocated if ndims > 1.
  Fix allocate and apply valid checks.
  Bugfix for clang compiler error on 3D Morton indexing. ATK-571.
  Add parameter checking on schema or dtype in view and group functions.
  Refactored and improved implementation of quest's Vector.
  Quest print functions avoid sending characters to ostream.
  ...

+++ src/components/shroud/tests/example/wrapExClass2.cpp
@@ -59,7 +59,7 @@ return rv.c_str();
 // splicer end class.ExClass2.method.get_name
 }
 
-// void getName(string_result_as_arg & SH_F_rv+intent(out)+len(LSH_F_rv)) const
+// void getName(string & SH_F_rv+intent(out)+len(LSH_F_rv)) const
 // function_index=36
 void AA_exclass2_get_name_bufferify(AA_exclass2 * self, char * SH_F_rv, int LSH_F_rv)
 {
@@ -136,7 +136,7 @@ int AA_exclass2_get_type_id(AA_exclass2 * self)
 ExClass2 *selfobj = static_cast<ExClass2 *>(static_cast<void *>(self));
 // splicer begin class.ExClass2.method.get_type_id
 TypeID rv = selfobj->getTypeID();
-return rv;
+return static_cast<int>(rv);
 // splicer end class.ExClass2.method.get_type_id
 }
